requirements.txt
bcrypt==3.2.0
blinker==1.4
certifi==2021.5.30
cffi==1.14.5
click==8.0.1
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.3.0
pycaret==2.3.1
pycparser==2.20
six==1.16.0
SQLAlchemy==1.4.20
Werkzeug==2.0.1
WTForms==2.3.3
